Validation pass 02: fix regressions + drop async waffle + tighten options
Second external-developer validation after v1.1. Found 8 items: 4
regressions I introduced in v1.1 when rewriting document.rs, plus 4
fresh findings. All fixed before closing the freeze.

Full report: xfa-program-office/SDK_CORE_FACADE_VALIDATION_PASS_02.md

Regressions fixed (missing methods from v1.1 rewrite)
-----------------------------------------------------
- `redact(text, RedactOptions)` restored on PdfDocument. Now takes the
  previously-declared-but-unused RedactOptions struct (was dead code).
- `redact_region(page, rect)` restored.
- `split_pages() -> Result<Vec<PdfDocument>>` restored. Required by
  the "split PDF" website how-to family.
- `extract_pages(range)` added (new; covers "split by page range"
  website how-to. Accepts any RangeBounds<usize>).
- `form_fields()` restored to `Result<Vec<FormField>>`. My v1.1 change
  to drop Result was wrong — form reads can fail on malformed dicts.

Fresh findings fixed
--------------------
- Async dropped entirely from 1.0. The `async-tokio` feature +
  `r#async` module shipped as an empty stub in v1.1, which is worse
  than either fully supporting or fully omitting. Deferred to 1.1
  with concrete scope. Removed feature, module, and tokio dep.
- `Permissions` now has `with_print / with_modify / with_copy /
  with_annotate / with_fill_forms / with_extract_accessibility /
  with_assemble / with_print_high_quality` builder methods. Previously
  users could only pick from 3 presets; custom combos were impossible.
- `SignOptions::visible_rect(page: u32, ...)` changed to
  `visible_rect(page: usize, ...)` for cross-module consistency.
  Every other page-index in the API is usize.
- `RedactOptions` now actually used by `redact()` signature. Added
  `on_pages(&[usize])` for page-scope redaction matching the
  website's "redact on specific pages" how-to.

RFC changes
-----------
- §13 updated: async-API entry rewritten to reflect full defer
  (not "beta in 1.0"). Added image-watermarks row pointing to #1225.
- §14 v1.2 revision log entry with full change list.

Tests
-----
5 passed, 5 ignored (all with linked Epic 2 issues), 0 failed.

Lesson learned — v1.1 regression cause
--------------------------------------
When I rewrote document.rs wholesale in validation pass 01, four
methods (redact, redact_region, split_pages, and form_fields Result)
silently disappeared. Pass 02 caught this by doing a fresh API
inventory independent of the v1.0 → v1.1 diff. Regression prevention
rule: after non-trivial API changes, enumerate ALL public methods
against the previous spec, not just check the diff.

Refs #1214 #1239 (2nd validation pass bewijs) #1240 (bootstrap suite intact)

+++ crates/pdfluent/src/signer.rs
@@ -34,7 +34,7 @@ pub struct SignOptions {
     pub(crate) location: Option<String>,
     pub(crate) contact_info: Option<String>,
     pub(crate) field_name: Option<String>,
-    pub(crate) visible_rect: Option<(u32, [f64; 4])>,
+    pub(crate) visible_rect: Option<(usize, [f64; 4])>,
     pub(crate) profile: PadesProfile,
 }
 
@@ -69,7 +69,7 @@ impl SignOptions {
     }
 
     /// Place a visible signature appearance on a specific page.
-    pub fn visible_rect(mut self, page: u32, rect: [f64; 4]) -> Self {
+    pub fn visible_rect(mut self, page: usize, rect: [f64; 4]) -> Self {
         self.visible_rect = Some((page, rect));
         self
     }
